requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
jinja2==3.1.2  # For HTML templates
aiofiles==23.2.1  # For async file operations
markdown==3.5.1  # For converting markdown to HTML/blocks
pyyaml==6.0.1  # For parsing YAML frontmatter

# AI/ML
anthropic==0.39.0
openai==1.51.0  # Updated to latest version for better async support
sentence-transformers==2.2.2
tiktoken==0.5.2

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
qdrant-client==1.7.0
pgvector==0.2.4
alembic==1.13.1
supabase==2.3.4  # For Supabase auth and database

# Text processing
textstat==0.7.3
markdown==3.5.1
beautifulsoup4==4.12.2
python-multipart==0.0.6
pyyaml==6.0.1
lxml==4.9.3  # For better HTML parsing with BeautifulSoup

# HTTP clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
python-dateutil==2.9.0.post0
pytz==2024.1

# GraphQL
gql[httpx]==3.5.0
graphql-core==3.2.3

# Utils
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0  # For JWT token handling
email-validator==2.1.0
tenacity==8.2.3  # For retry logic
cryptography==41.0.7  # For API key encryption

# API Protection
slowapi==0.1.9  # Rate limiting for FastAPI

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.0
isort==5.13.2